secure_platform/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "secure_platform"
version = "0.1.0"
description = "Secure Data Platform with Zero Trust architecture, mTLS, RBAC/ABAC, and audit logs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
dependencies = [
    "pydantic>=2.0",
    "structlog>=24.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "prometheus-client>=0.19",
    "cryptography>=42.0",
    "pycasbin>=1.35",
    "grpcio>=1.60",
    "pyjwt>=2.8",
    "python-jose>=3.3",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "hypothesis>=6.98",
    "pytest-cov>=4.1",
    "mypy>=1.8",
    "ruff>=0.3",
    "black>=24.2",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]

[tool.black]
line-length = 100

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
